# Trucking-Game-API/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Trucking Software Public API",
		"description": "API for trucking software web application",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:3001/api/v1",
			"description": "Development server"
		}
	],
	"paths": {
...
		"/categories": {
			"get": {
				"description": "Get a paginated list of categories",
				"summary": "Get a paginated list of categories",
				"operationId": "getAllCategories",
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "Number of documents to return. Defaults to 10."
					},
					{
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "Page number used for determining number of documents to skip. Defaults to 1."
					}
				],
				"tags": ["categories"],
				"responses": {
					"200": {
						"description": "Get paginated categories success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"categories": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Category"
											}
										},
										"totalDocuments": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get paginated categories fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"post": {
				"description": "Create a new category",
				"summary": "Create a new category",
				"operationId": "createCategory",
				"tags": ["categories"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryBody"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Create category success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"category": {
											"$ref": "#/components/schemas/Category"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/categories/options": {
			"get": {
				"description": "Get a list of all categories",
				"summary": "Get a list of all categories",
				"operationId": "getCategoryOptions",
				"tags": ["categories"],
				"responses": {
					"200": {
						"description": "Get all categories success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "string",
											"example": "true"
										},
										"categories": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Category"
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get all categories fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/categories/search": {
			"get": {
				"description": "Get a list of articles that belong to 1 or more categories",
				"summary": "Get a list of articles by categories",
				"operationId": "getArticlesByCategoryQuery",
				"tags": ["categories"],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						},
						"required": true,
						"description": "One or more category ids to search by"
					},
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The limit of documents to return. Defaults to 10."
					},
					{
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The page of documents to return, will determine the amount of documents to skip when querying. Defaults to 1."
					}
				],
				"responses": {
					"200": {
						"description": "Query articles by category success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"articles": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Article"
											}
										},
										"totalDocuments": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Query articles by category fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/categories/{categoryId}": {
			"parameters": [
				{
					"in": "path",
					"name": "categoryId",
					"schema": {
						"type": "string"
					},
					"required": true,
					"description": "Id of a specific category"
				}
			],
			"get": {
				"description": "Get a single category by id",
				"summary": "Get a single category by id",
				"operationId": "getCategoryById",
				"tags": ["categories"],
				"responses": {
					"200": {
						"description": "Get category by id success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"category": {
											"$ref": "#/components/schemas/Category"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get category by id fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"delete": {
				"description": "Delete a single category by id",
				"summary": "Delete a single category by id",
				"operationId": "deleteOneCategory",
				"tags": ["categories"],
				"responses": {
					"200": {
						"description": "Delete category by id success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"message": {
											"type": "string",
											"example": "Successfully deleted category"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Delete category by id fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"patch": {
				"description": "Update a single category by id",
				"summary": "Update a single category by id",
				"operationId": "updateCategory",
				"tags": ["categories"],
				"responses": {
					"200": {
						"description": "Update category by id success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"updatedCategory": {
											"$ref": "#/components/schemas/Category"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Update category by id fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/categories/multiple": {
			"delete": {
				"description": "Delete multiple categories by id",
				"summary": "Delete multiple categories by id",
				"operationId": "deleteMultipleCategories",
				"tags": ["categories"],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						},
						"required": true,
						"description": "One or more ids of categories to be deleted"
					}
				],
				"responses": {
					"200": {
						"description": "Delete multiple categories success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"message": {
											"type": "string",
											"example": "Successfully deleted categories"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Delete multiple categories fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteMultipleErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/payments/sales": {
			"get": {
				"description": "Get sales graph data for past 5 minutes",
				"summary": "Get sales graph data for past 5 minutes",
				"operationId": "getSales",
				"tags": ["payments"],
				"responses": {
					"200": {
						"description": "Get sales graph data success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"label": {
											"type": "string"
										},
										"data": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get sales graph data fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/payments/create-checkout-session": {
			"post": {
				"description": "Create a stripe checkout session",
				"summary": "Create a stripe checkout session",
				"operationId": "createCheckoutSession",
				"tags": ["payments"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"successUrl": {
										"type": "string",
										"example": "/receipt"
									},
									"cancelUrl": {
										"type": "string",
										"example": "/home"
									}
								},
								"required": ["successUrl", "cancelUrl"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Create checkout session success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"id": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Create checkout session fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
			"Article": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/User"
					},
					"slug": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"htmlBody": {
						"type": "string"
					},
					"textBody": {
						"type": "string"
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"comments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArticleComment"
						}
					},
					"isFree": {
						"type": "boolean"
					},
					"editingLocked": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
...
			"ArticleComment": {
				"type": "object",
				"properties": {
					"text": {
						"type": "string"
					},
					"relatedArticle": {
						"$ref": "#/components/schemas/Article"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
...
			"Category": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
			"CategoryBody": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": ["name", "description"]
			},
			"DeleteMultipleErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": "false"
					},
					"message": {
						"type": "string",
						"example": "One or more artcles was not found"
					},
					"incomplete": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": "false"
					},
					"error": {
						"type": "string"
					}
				}
			},
...
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "integer"
					},
					"phone": {
						"type": "string"
					},
					"profileImage": {
						"type": "string"
					},
					"hasPaid": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
...
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	]
}